g4: dtsi: massive thermal handling changes (1/2)
... part 2/2 happens in the device tree (g4-common, thermal-engine).

TL;DR:

The kernel handles hotplug of all cores if needed (just HIGHLY optimized now),
while frequencies will be handled by the resurrected thermal-engine service.

Details:

 - generally re-plug an unplugged core MUCH earlier
 - reducing the "early-boot" hotplug from 105 C to 90 C with clearance at 85 C
 - increasing hotplug of the little cluster from 90 C to 100 C with clearance at 90 C
 - increasing hotplug of the big cluster from 90 C to 98 C with clearance at 70 C
 - stop using frequency handling by the kernel
 - increasing polling interval from 25 ms to 50 ms

The new hotplug handling approach:

Before we had situations where unplug/plug happens too early/late which causes
performance slow-downs and other issues. This is fixed now by unplugging later with a
much sooner clearance (i.e. re-plug a core).
Unplugging a core is the most effective way to reduce heat obviously but ofc this
comes with dramatic performance losses.
The way it is integrated now ensures the device does not get to hot (105 actually IS
too hot, which was the default by LG) while bringing a core back much earlier into
the game.

The new frequency handling approach:

Frequencies on kernel level are not very flexible and so you need the thermal-engine
service for monitoring and handling that correctly. Actually the frequency handling here
has been "disabled" (i.e. using an impossible high value) to avoid funny ping-pong
games with the user-space service.
Other then that without a thermal service running (which was the case until now!!!)
the GPU and battery are not monitored and so can cause damages to the device. So
resurrecting that service is needed in any case, IMHO.

Signed-off-by: steadfasterX <steadfasterX@gmail.com>